requirements.txt
discord.py
flask
matplotlib
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
gspread